refactor: rename session-account.ts to account.ts and outside-execution.ts to execution.ts
The session/ts/ directory already establishes context, making the
prefixes redundant.
