+++ packages/controller/src/session/account.ts
@@ -1,5 +1,5 @@
 import type { Policy } from "./ts/types";
-import { CartridgeSessionAccount } from "./ts/session-account";
+import { CartridgeSessionAccount } from "./ts/account";
 import { Call, InvokeFunctionResponse, WalletAccount } from "starknet";
 
 import { normalizeCalls } from "../utils";
