+++ packages/controller/src/session/ts/index.ts
@@ -2,7 +2,7 @@ export * from "./types";
 export * from "./errors";
 export * from "./guid";
 export * from "./merkle";
-export * from "./outside-execution";
-export * from "./session-account";
+export * from "./execution";
+export * from "./account";
 export * from "./subscribe";
 export * from "./shared";
